@@ -15,14 +15,17 @@
     "@types/express": "^4.0.39",
     "@types/mongodb": "^3.0.15",
     "@types/mongoose": "^4.7.24",
-    "@types/node": "^8.0.47",
+    "@types/mongoose-auto-increment": "^5.0.30",
+    "@types/mongoose-sequence": "^3.0.3",
     "@types/request": "^2.0.7",
     "nodemon": "^1.12.1",
     "ts-node": "^3.3.0",
-    "typescript": "^2.5.3"
+    "typescript": "^3.8.3"
   },
   "dependencies": {
     "@types/body-parser": "^1.16.8",
+    "@types/log4js": "^2.3.5",
+    "@types/node": "^13.9.8",
     "@types/mongoose-geojson-schema": "^2.1.1",
     "@types/multer": "^1.3.6",
     "@types/q": "^1.0.7",
@@ -32,8 +35,9 @@
     "circle-to-polygon": "^1.0.2",
     "cors": "^2.8.4",
     "express": "^4.16.2",
+    "log4js": "^6.3.0",
     "mongodb": "^3.0.7",
-    "mongoose": "^4.12.5",
+    "mongoose": "^5.11.5",
     "mongoose-geojson-schema": "^2.1.2",
     "multer": "^1.3.0",
     "q": "^1.5.1",
