@@ -3,7 +3,9 @@ import { Storage } from '@ionic/storage';
 import { GeocoderProvider } from '../../providers/geocoder/geocoder';
 import { NavController , AlertController } from 'ionic-angular';
 import { WebServiceProvider } from '../../providers/web-service/web-service';
-
+import { Logger, LogLevel } from 'ask-logger';
+const LOGGER = Logger.getLogger('HelloIonicPage')
+LOGGER.set_level(LogLevel.DEBUG)
 @Component({
   selector: 'page-hello-ionic',
   templateUrl: 'hello-ionic.html'
@@ -22,14 +24,13 @@ export class HelloIonicPage {
 
   wfs() {
     this.rest.getWFS().subscribe( (val) => {
-      console.log("RISULTATO",val);
+      LOGGER.info("[WFS] resutl ",val);
     })
   }
 
   verifica() {
     this.storage.get('utente').then((val) => {
-      console.log(val);
-      console.log(val.email);
+      LOGGER.info("[verifica] resutl ", val);
     });
   }
 
