@@ -11,6 +11,10 @@ import { WebServiceProvider } from '../../providers/web-service/web-service';
 import { Utente } from '../../models/Utente';
 import { IscrizionePage } from '../iscrizione/iscrizione';
 import { Storage } from '@ionic/storage';
+
+import { Logger, LogLevel } from 'ask-logger';
+const LOGGER = Logger.getLogger('TakepicturePage')
+LOGGER.set_level(LogLevel.DEBUG)
 @Component({
   selector: 'page-takepicture',
   templateUrl: 'takepicture.html',
@@ -46,23 +50,23 @@ export class TakepicturePage {
     public toastCtrl: ToastController,
     private camera: Camera,
     public plt: Platform,
-    private storage: Storage, 
+    private storage: Storage,
     public rest: WebServiceProvider,
     private file: File) {
 
     if (this.plt.is('core') || this.plt.is('mobileweb')) {
-      console.log('È UN BROWSER');
+      LOGGER.info("[constructor] IS A BROWSER");
       this.isApp = false;
 
 
     } else {
-      console.log('NON È UN BROWSER');
+      LOGGER.info("[constructor] IS NOT A BROWSER");
       this.isApp = true;
 
     }
 
     this.storage.get('utente').then((val) => {
-      console.log(val);
+
       this.utente = val;
       if (val == null) {
         let alert = this.alertCtrl.create({
@@ -87,79 +91,26 @@ export class TakepicturePage {
   private photo_split: String;
   private nativeFilePath: String;
 
-  // options: CameraOptions = {
-  //   quality: 100,
-  //   saveToPhotoAlbum: false,
-  //   correctOrientation: true,
-  //   destinationType: this.camera.DestinationType.DATA_URL,
-  //   encodingType: this.camera.EncodingType.JPEG,
-  //   mediaType: this.camera.MediaType.PICTURE
-  // }
-
-
   getImage() {
-
-
-
-
     const options: CameraOptions = {
       quality: 100,
       destinationType: this.camera.DestinationType.FILE_URI,
       sourceType: this.camera.PictureSourceType.PHOTOLIBRARY,
-      // allowEdit: true,
-      // saveToPhotoAlbum: false,
-      // correctOrientation: true
     }
 
     this.camera.getPicture(options).then((imageData) => {
+      LOGGER.info("[getImage] imagedata",imageData);
+      LOGGER.info("[getImage] doomsanitaizer",this.domSanitizer.bypassSecurityTrustResourceUrl(imageData));
 
-      console.log("--------------------------------");
-      console.log(imageData);
-      console.log(this.domSanitizer.bypassSecurityTrustResourceUrl(imageData));
-      console.log("--------------------------------");
       this.nativeFilePath = imageData;
       this.imageOldPath = imageData;
 
       this.filePath.resolveNativePath(imageData)
         .then(filePath => {
-          console.log("--------------------------------");
-
           this.imageFileName = filePath;
-          // console.log(this.imageURI);
-          // console.log("-------------**-------------------");
-
-          // let sourceDirectory = this.imageURI.substring(0, this.imageURI.lastIndexOf('/') + 1);
-          // let sourceFileName = this.imageURI.substring(this.imageURI.lastIndexOf('/') + 1, this.imageURI.length);
-          // sourceFileName = sourceFileName.split('?').shift();
-          // this.file.copyFile(sourceDirectory, sourceFileName, this.file.externalApplicationStorageDirectory, sourceFileName).then((result: any) => {
-
-
-          //   this.imageNewPath = result.nativeURL;
-
-          //   console.log("-***************-------------------****************-");
-          //   console.log(this.imageNewPath);
-          //   console.log("-***************-------------------****************-");
-
-          //   }, (err) => {
-          //     alert(JSON.stringify(err));
-          //   })
         });
-
-      // if (imageData.substring(0, 21) == "content://com.android") {
-      // this.photo_split = imageData.split("%3A");
-      // imageData = imageData.replace("%", "%25");
-      //   this.imageFileName = "content://media/external/images/media/" + this.photo_split[1];
-      //   this.imageURI = "content://media/external/images/media/" + this.photo_split[1];
-      // } else {
-
-      // }
-
-
-
-
     }, (err) => {
-      console.log(err);
-
+      LOGGER.error("[getImage]", err);
       this.presentToast(err);
 
     });
@@ -167,22 +118,17 @@ export class TakepicturePage {
   }
 
   getFile(event) {
-    console.log("-|->", JSON.stringify(event));
-    // let  file : File = $event.target.files[0];
-    // console.log("-|2->", JSON.stringify(file));
-
-
-    console.log("-|3->", JSON.stringify(this.filesName));
-    console.log("-|3->", this.filesName.nativeElement);
+    LOGGER.info("[getFile] event ",event);
+    LOGGER.info("[getFile] this.filesName.nativeElement ", this.filesName.nativeElement);
     this.imageOldPath = this.filesName.nativeElement.value;
     this.toUpload = this.filesName.nativeElement.files[0];
-    console.log("-|4->", this.filesName.nativeElement.files.length);
-    console.log("-|4->", this.filesName.nativeElement.files[0].name);
-    // console.log(this.pic);
+    LOGGER.info("[getFile] this.filesName.nativeElement.files.length ", this.filesName.nativeElement.files.length);
+    LOGGER.info("[getFile] this.filesName.nativeElement.files[0].name ", this.filesName.nativeElement.files[0].name);
+
   }
 
   uploadFile() {
-    console.log("-->", this.filepathb);
+    LOGGER.info("[uploadFile]this.filepathb", this.filepathb);
     if (this.imageOldPath == null) {
       let alert = this.alertCtrl.create({
         title: 'Attenzione',
@@ -197,9 +143,7 @@ export class TakepicturePage {
       loader.present();
 
       if (this.isApp) {
-        console.log("--------------------------------------------------------------------");
-        console.log("E' un app");
-        console.log("--------------------------------------------------------------------");
+        LOGGER.info("[uploadFile] IS A APP");
         const fileTransfer: FileTransferObject = this.transfer.create();
 
         let options: FileUploadOptions = {
@@ -212,35 +156,30 @@ export class TakepicturePage {
             formModel: this.navParams.get("formmodel")
           }
         }
-        console.log("-*******************************-");
-        console.log(this.imageOldPath);
-        console.log("-*******************************-");
-        // fileTransfer.upload(this.imageOldPath, 'http://192.168.1.8:3000/offerta/upload', options)
+
+        LOGGER.info("[uploadFile]this.imageOldPath:",this.imageOldPath);
+
+
         fileTransfer.upload(this.imageOldPath, WebServiceProvider.UPLOADURL, options)
           .then((data) => {
 
-            console.log("Code = " + data.responseCode)
-            console.log("Response = " + data.response)
-            console.log("Sent = " + data.bytesSent)
-            console.log(data + " Uploaded Successfully");
-            console.log("->", data);
-            // this.imageFileName = "http://192.168.0.7:8080/static/images/ionicfile.jpg"
+            LOGGER.info("[uploadFile]Code = " + data.responseCode)
+            LOGGER.info("[uploadFile]Response = " + data.response)
+            LOGGER.info("[uploadFile]Sent = " + data.bytesSent)
+            LOGGER.info("[uploadFile]data",data," Uploaded Successfully");
             loader.dismiss();
             this.presentToast("Annuncio caricato!");
             this.navCtrl.setRoot(HelloIonicPage);
-            // this.navCtrl.push();
+
           }, (err) => {
             console.log(err);
             loader.dismiss();
             this.presentToast(err);
           });
       } else {
-
-        console.log("--------------------------------------------------------------------");
-        console.log("NON E' un app  ++++++++++++++++++");
-        console.log("--------------------------------------------------------------------");
+        LOGGER.info("[uploadFile] IS A BROWSWER");
           this.rest.postFile(this.toUpload, this.navParams.get("formmodel"),this.utente ).subscribe( (data:Risposta) =>{
-            console.log("----Risposta--->",JSON.stringify(data));
+            LOGGER.info("[uploadFile]DATA",data);
             loader.dismiss();
             if (data.esito) {
               this.presentToast("Annuncio caricato!");
@@ -248,9 +187,9 @@ export class TakepicturePage {
             } else {
               this.presentToast(data.status);
             }
-            
+
           });
-         
+
       }
     }
   }
@@ -270,36 +209,4 @@ export class TakepicturePage {
   }
 
 }
-
-  // onTakePicture() {
-  //   const options: CameraOptions = {
-  //     quality: 100,
-  //     destinationType: this.camera.DestinationType.DATA_URL,
-  //     saveToPhotoAlbum: true,
-  //     mediaType: this.camera.MediaType.PICTURE
-  //   }
-
-
-  //   console.log(JSON.stringify(this.navParams.get("formmodel")));
-  //   this.camera.getPicture(options).then((imageData) => {
-  //     this.image = 'data:image/jpeg;base64,' + imageData;
-  //   }, (err) => {
-  //     this.displayErrorAlert(err);
-  //   });
-  // }
-
-  // displayErrorAlert(err) {
-  //   console.log(err);
-  //   let alert = this.alertCtrl.create({
-  //     title: 'Error',
-  //     subTitle: 'Error while trying to capture picture',
-  //     buttons: ['OK']
-  //   });
-  //   alert.present();
-  // }
-
-
-
-
-
 
